Update AzureML step operator to use SDK v2 and service connectors
