+++ src/zenml/integrations/azure/orchestrators/azureml_orchestrator.py
@@ -40,15 +40,15 @@ from azure.ai.ml.entities import (
 from azure.core.exceptions import (
     HttpResponseError,
     ResourceExistsError,
-    ResourceNotFoundError,
 )
 from azure.identity import DefaultAzureCredential
 
 from zenml.config.base_settings import BaseSettings
 from zenml.config.step_configurations import Step
 from zenml.enums import StackComponentType
+from zenml.integrations.azure.azureml_utils import create_or_get_compute
+from zenml.integrations.azure.flavors.azureml import AzureMLComputeTypes
 from zenml.integrations.azure.flavors.azureml_orchestrator_flavor import (
-    AzureMLComputeTypes,
     AzureMLOrchestratorConfig,
     AzureMLOrchestratorSettings,
 )
@@ -62,8 +62,6 @@ from zenml.stack import StackValidator
 from zenml.utils.string_utils import b64_encode
 
 if TYPE_CHECKING:
-    from azure.ai.ml.entities import AmlCompute, ComputeInstance
-
     from zenml.models import PipelineDeploymentResponse
     from zenml.stack import Stack
 
@@ -193,172 +191,6 @@ class AzureMLOrchestrator(ContainerizedOrchestrator):
             command=" ".join(command + arguments),
         )
 
-    @staticmethod
-    def _check_settings_and_compute_configuration(
-        parameter: str,
-        settings: AzureMLOrchestratorSettings,
-        compute: Union["ComputeInstance", "AmlCompute"],
-    ) -> None:
-        """Utility function comparing a parameter between settings and compute.
-
-        Args:
-            parameter: the name of the parameter.
-            settings: The AzureML orchestrator settings.
-            compute: The compute instance or cluster from AzureML.
-        """
-        # Check the compute size
-        compute_value = getattr(compute, parameter)
-        settings_value = getattr(settings, parameter)
-
-        if settings_value is not None and settings_value != compute_value:
-            logger.warning(
-                f"The '{parameter}' defined in the settings '{settings_value}' "
-                "does not match the actual parameter of the instance: "
-                f"'{compute_value}'. Will ignore this setting for now."
-            )
-
-    def _create_or_get_compute(
-        self, client: MLClient, settings: AzureMLOrchestratorSettings
-    ) -> Optional[str]:
-        """Creates or fetches the compute target if defined in the settings.
-
-        Args:
-            client: the AzureML client.
-            settings: the settings for the orchestrator.
-
-        Returns:
-            None, if the orchestrator is using serverless compute or
-            str, the name of the compute target (instance or cluster).
-
-        Raises:
-            RuntimeError: if the fetched compute target is unsupported or the
-                mode defined in the setting does not match the type of the
-                compute target.
-        """
-        # If the mode is serverless, we can not fetch anything anyhow
-        if settings.mode == AzureMLComputeTypes.SERVERLESS:
-            return None
-
-        # If a name is not provided, generate one based on the orchestrator id
-        compute_name = settings.compute_name or f"compute_{self.id}"
-        # Try to fetch the compute target
-        try:
-            compute = client.compute.get(compute_name)
-
-            logger.info(f"Using existing compute target: '{compute_name}'.")
-
-            # Check if compute size matches with the settings
-            self._check_settings_and_compute_configuration(
-                parameter="size", settings=settings, compute=compute
-            )
-
-            compute_type = compute.type
-
-            # Check the type and matches the settings
-            if compute_type == "computeinstance":  # Compute Instance
-                if settings.mode != AzureMLComputeTypes.COMPUTE_INSTANCE:
-                    raise RuntimeError(
-                        "The mode of operation for the compute target defined"
-                        f"in the settings '{settings.mode}' does not match "
-                        f"the type of the compute target: `{compute_name}` "
-                        "which is a 'compute-instance'. Please make sure that "
-                        "the settings are adjusted properly."
-                    )
-
-                if compute.state != "Running":
-                    raise RuntimeError(
-                        f"The compute instance `{compute_name}` is not in a "
-                        "running state at the moment. Please make sure that "
-                        "the compute target is running, before executing the "
-                        "pipeline."
-                    )
-
-                # Idle time before shutdown
-                self._check_settings_and_compute_configuration(
-                    parameter="idle_time_before_shutdown_minutes",
-                    settings=settings,
-                    compute=compute,
-                )
-
-            elif compute_type == "amIcompute":  # Compute Cluster
-                if settings.mode != AzureMLComputeTypes.COMPUTE_CLUSTER:
-                    raise RuntimeError(
-                        "The mode of operation for the compute target defined "
-                        f"in the settings '{settings.mode}' does not match "
-                        f"the type of the compute target: `{compute_name}` "
-                        "which is a 'compute-cluster'. Please make sure that "
-                        "the settings are adjusted properly."
-                    )
-
-                if compute.provisioning_state != "Succeeded":
-                    raise RuntimeError(
-                        f"The provisioning state '{compute.provisioning_state}'"
-                        f"of the compute cluster `{compute_name}` is not "
-                        "successful. Please make sure that the compute cluster "
-                        "is provisioned properly, before executing the "
-                        "pipeline."
-                    )
-
-                for parameter in [
-                    "idle_time_before_scale_down",
-                    "max_instances",
-                    "min_instances",
-                    "tier",
-                    "location",
-                ]:
-                    # Check all possible configurations
-                    self._check_settings_and_compute_configuration(
-                        parameter=parameter, settings=settings, compute=compute
-                    )
-            else:
-                raise RuntimeError(f"Unsupported compute type: {compute_type}")
-            return compute_name
-
-        # If the compute target does not exist create it
-        except ResourceNotFoundError:
-            logger.info(
-                "Can not find the compute target with name: "
-                f"'{compute_name}':"
-            )
-
-            if settings.mode == AzureMLComputeTypes.COMPUTE_INSTANCE:
-                logger.info(
-                    "Creating a new compute instance. This might take a "
-                    "few minutes."
-                )
-
-                from azure.ai.ml.entities import ComputeInstance
-
-                compute_instance = ComputeInstance(
-                    name=compute_name,
-                    size=settings.size,
-                    idle_time_before_shutdown_minutes=settings.idle_time_before_shutdown_minutes,
-                )
-                client.begin_create_or_update(compute_instance).result()
-                return compute_name
-
-            elif settings.mode == AzureMLComputeTypes.COMPUTE_CLUSTER:
-                logger.info(
-                    "Creating a new compute cluster. This might take a "
-                    "few minutes."
-                )
-
-                from azure.ai.ml.entities import AmlCompute
-
-                compute_cluster = AmlCompute(
-                    name=compute_name,
-                    size=settings.size,
-                    location=settings.location,
-                    min_instances=settings.min_instances,
-                    max_instances=settings.max_instances,
-                    idle_time_before_scale_down=settings.idle_time_before_scaledown_down,
-                    tier=settings.tier,
-                )
-                client.begin_create_or_update(compute_cluster).result()
-                return compute_name
-
-        return None
-
     def prepare_or_run_pipeline(
         self,
         deployment: "PipelineDeploymentResponse",
@@ -429,7 +261,9 @@ class AzureMLOrchestrator(ContainerizedOrchestrator):
         )
         pipeline_args["name"] = run_name
 
-        if compute_target := self._create_or_get_compute(ml_client, settings):
+        if compute_target := create_or_get_compute(
+            ml_client, settings, default_compute_name=f"zenml_{self.id}"
+        ):
             pipeline_args["compute"] = compute_target
 
         @pipeline(force_rerun=True, **pipeline_args)  # type: ignore[call-overload, misc]
@@ -540,5 +374,17 @@ class AzureMLOrchestrator(ContainerizedOrchestrator):
                 )
 
         else:
-            ml_client.jobs.create_or_update(pipeline_job)
+            job = ml_client.jobs.create_or_update(pipeline_job)
             logger.info(f"Pipeline {run_name} has been started.")
+            assert job.services is not None
+            assert job.name is not None
+
+            logger.info(
+                f"Pipeline {run_name} is running. "
+                "You can view the pipeline in the AzureML portal at "
+                f"{job.services['Studio'].endpoint}"
+            )
+
+            if self.config.synchronous:
+                logger.info("Waiting for pipeline to finish...")
+                ml_client.jobs.stream(job.name)
